Ошибки logger до конца переведены на Английский язык

+++ getChange.py
@@ -102,6 +102,10 @@ def GetChange(ChangeLinesIndex: list[tuple[Any]], OldCode: str, NewCode: str) ->
                 ChangeLine = OldCode.splitlines(keepends=True)[OldLineIndex]
                 if not ChangeLine.isspace():
                     ChangeLine = ChangeLine.replace('\n','')
+
+                if ChangeLine.isspace() or ChangeLine.strip() == "" and ChangeType == 'del':
+                    continue
+
                 IndexDict = FindStartEndPositionSubStringInStr(OldCode, ChangeLine, OldLineIndex)
                 Changes.append({ 'type': 'delete', 'start': IndexDict["StartIndex"], 'end': IndexDict["EndIndex"], 'added': '', 'deleted': ChangeLine, "NewLineIndex": NewLineIndex})
             else:
